+++ backend/src/main/java/ch/cevi/db/adapter/domain/CeviEvent.java
@@ -1,11 +1,12 @@
 package ch.cevi.db.adapter.domain;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Objects;
 
 public record CeviEvent(String id, String name, String description, String applicationLink, LocalDateTime startsAt,
                         LocalDateTime finishAt, String group, String location, String kind, CeviEventType eventType,
-                        int participantsCount, Integer maximumParticipants) {
+                        int participantsCount, Integer maximumParticipants, LocalDate applicationOpeningAt, LocalDate applicationClosingAt) {
     public CeviEvent {
         Objects.requireNonNull(id);
         Objects.requireNonNull(name, () -> "Name mustn't be null for event " + id);
@@ -22,8 +23,20 @@ public record CeviEvent(String id, String name, String description, String appli
         }
     }
 
-    public CeviEvent withMaximumParticipants(int maximumParticipants) {
-        return new CeviEvent(id, name, description, applicationLink, startsAt, finishAt, group, location, kind, eventType, participantsCount, maximumParticipants);
+    public CeviEvent withParticipantsCount(Integer participantsCount) {
+        return new CeviEvent(id, name, description, applicationLink, startsAt, finishAt, group, location, kind, eventType, participantsCount, maximumParticipants, applicationOpeningAt, applicationClosingAt);
+    }
+
+    public CeviEvent withMaximumParticipants(Integer maximumParticipants) {
+        return new CeviEvent(id, name, description, applicationLink, startsAt, finishAt, group, location, kind, eventType, participantsCount, maximumParticipants, applicationOpeningAt, applicationClosingAt);
+    }
+
+    public CeviEvent withApplicationOpeningAt(LocalDate applicationOpeningAt) {
+        return new CeviEvent(id, name, description, applicationLink, startsAt, finishAt, group, location, kind, eventType, participantsCount, maximumParticipants, applicationOpeningAt, applicationClosingAt);
+    }
+
+    public CeviEvent withApplicationClosingAt(LocalDate applicationClosingAt) {
+        return new CeviEvent(id, name, description, applicationLink, startsAt, finishAt, group, location, kind, eventType, participantsCount, maximumParticipants, applicationOpeningAt, applicationClosingAt);
     }
 
     public boolean hasLimitedCapacity() {
@@ -32,5 +45,10 @@ public record CeviEvent(String id, String name, String description, String appli
 
     public boolean hasAvailablePlaces() {
         return this.maximumParticipants() == null || this.maximumParticipants() > this.participantsCount();
+    }
+
+    public boolean isApplicationOpen() {
+        return (this.applicationClosingAt == null || LocalDate.now().isBefore(this.applicationClosingAt) || LocalDate.now().isEqual(this.applicationClosingAt)) &&
+                (this.applicationOpeningAt == null || LocalDate.now().isAfter(this.applicationOpeningAt) || LocalDate.now().isEqual(this.applicationOpeningAt));
     }
 }