+++ backend/src/main/java/ch/cevi/db/adapter/hitobito/HitobitoEvent.java
@@ -1,9 +1,10 @@
 package ch.cevi.db.adapter.hitobito;
 
+import java.time.LocalDate;
 import java.util.Objects;
 
 public record HitobitoEvent(String id, String name, String description, String external_application_link, HitobitoLinks links,
-                            int participant_count, Integer maximum_participants) {
+                            int participant_count, Integer maximum_participants, LocalDate application_opening_at, LocalDate application_closing_at) {
     public HitobitoEvent {
         Objects.requireNonNull(id);
         Objects.requireNonNull(name);