+++ backend/src/test/java/ch/cevi/db/adapter/EventFilterTest.java
@@ -3,6 +3,8 @@ package ch.cevi.db.adapter;
 import ch.cevi.db.adapter.fixtures.CeviEventFixture;
 import org.junit.jupiter.api.Test;
 
+import java.time.LocalDate;
+
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 public class EventFilterTest {
@@ -40,5 +42,18 @@ public class EventFilterTest {
         var eventOverbooked = ceviEvent.withMaximumParticipants(ceviEvent.participantsCount());
         // match as there are already too many participants
         assertThat(sut.match(eventOverbooked)).isTrue();
+    }
+
+    @Test
+    void should_filter_open_application() {
+        var sut = EventFilter.emptyFilter().withIsApplicationOpen(true);
+
+        var applicationNotYetOpen = CeviEventFixture.ceviEvent()
+                .withApplicationOpeningAt(LocalDate.now().plusDays(1));
+        assertThat(sut.match(applicationNotYetOpen)).isFalse();
+
+        var applicationOpen = CeviEventFixture.ceviEvent()
+                .withApplicationOpeningAt(LocalDate.now().minusDays(1));
+        assertThat(sut.match(applicationOpen)).isTrue();
     }
 }