+++ backend/src/test/java/ch/cevi/db/adapter/fixtures/CeviEventFixture.java
@@ -19,6 +19,8 @@ public class CeviEventFixture {
                 "",
                 CeviEventType.EVENT,
                 3,
+                null,
+                null,
                 null);
     }
 }